+++ .changeset/config.json
@@ -23,7 +23,6 @@
       "@remirror/dom",
       "@remirror/extension-annotation",
       "@remirror/extension-auto-link",
-      "@remirror/extension-base-keymap",
       "@remirror/extension-bidi",
       "@remirror/extension-blockquote",
       "@remirror/extension-bold",
