+++ .changeset/fluffy-peas-search.md
@@ -1,0 +1,5 @@
+---
+'@remirror/extension-base-keymap': major
+---
+
+Deprecated - this functionality is now one of the `builtin` extensions. Everything should work as before if you're on the latest version of `remirror`.
