+++ .changeset/pre.json
@@ -11,7 +11,6 @@
     "@remirror/dev": "0.13.2",
     "@remirror/dom": "0.0.0",
     "@remirror/extension-auto-link": "0.0.0",
-    "@remirror/extension-base-keymap": "0.0.0",
     "@remirror/extension-bidi": "0.0.0",
     "@remirror/extension-blockquote": "0.0.0",
     "@remirror/extension-bold": "0.0.0",
