+++ .changeset/twelve-mice-look.md
@@ -1,0 +1,7 @@
+---
+'remirror': major
+'@remirror/core': minor
+'@remirror/preset-core': patch
+---
+
+Move all keymap functionality to `KeymapExtension` from `@remirror/core`. Remove all references to `@remirror/extension-base-keymap`.
