--- packages/@remirror/extension-base-keymap/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@remirror/extension-base-keymap",
-  "version": "1.0.0-next.15",
-  "description": "Add familiar keyboard shortcuts to your beautiful remirror editor.",
-  "keywords": [
-    "remirror",
-    "extension",
-    "keymaps",
-    "keyboard",
-    "shortcut",
-    "productivity"
-  ],
-  "homepage": "https://github.com/remirror/remirror/tree/HEAD/packages/@remirror/extension-base-keymap",
-  "repository": "https://github.com/remirror/remirror/tree/HEAD/packages/@remirror/extension-base-keymap",
-  "license": "MIT",
-  "contributors": [
-    "Ifiok Jr. <ifiokotung@gmail.com>"
-  ],
-  "sideEffects": false,
-  "main": "dist/extension-base-keymap.cjs.js",
-  "module": "dist/extension-base-keymap.esm.js",
-  "types": "dist/extension-base-keymap.cjs.d.ts",
-  "files": [
-    "dist"
-  ],
-  "dependencies": {
-    "@babel/runtime": "^7.10.5"
-  },
-  "devDependencies": {
-    "@remirror/core": "^1.0.0-next.15",
-    "@remirror/pm": "1.0.0-next.4"
-  },
-  "peerDependencies": {
-    "@remirror/core": "^1.0.0-next.15",
-    "@remirror/pm": "1.0.0-next.4"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "meta": {
-    "sizeLimit": "5 KB"
-  }
-}
