--- packages/@remirror/extension-base-keymap/readme.md
@@ -1,43 +1,0 @@
-# @remirror/extension-base-keymap
-
-> Add familiar keyboard shortcuts to your beautiful remirror editor.
-
-[![Version][version]][npm] [![Weekly Downloads][downloads-badge]][npm]
-[![Bundled size][size-badge]][size] [![Typed Codebase][typescript]](./src/index.ts)
-![MIT License][license]
-
-[version]: https://flat.badgen.net/npm/v/@remirror/extension-base-keymap
-[npm]: https://npmjs.com/package/@remirror/extension-base-keymap
-[license]: https://flat.badgen.net/badge/license/MIT/purple
-[size]: https://bundlephobia.com/result?p=@remirror/extension-base-keymap
-[size-badge]: https://flat.badgen.net/bundlephobia/minzip/@remirror/extension-base-keymap
-[typescript]: https://flat.badgen.net/badge/icon/TypeScript?icon=typescript&label
-[downloads-badge]: https://badgen.net/npm/dw/@remirror/extension-base-keymap/red?icon=npm
-
-<br />
-
-## Installation
-
-```bash
-yarn add @remirror/extension-base-keymap@next @remirror/pm@next # yarn
-pnpm add @remirror/extension-base-keymap@next @remirror/pm@next # pnpm
-npm install @remirror/extension-base-keymap@next @remirror/pm@next # npm
-```
-
-<br />
-
-## Usage
-
-This is one of the **core** extensions and is included by default via `@remirror/preset-core`. It is
-responsible for adding most of the basic keyboard shortcuts you're familiar with when editing
-documents.
-
-`Backspace`, `Enter`, `Delete`, Select All (`cmd-a`) and a few others.
-
-<br />
-
-## Credits
-
-This package was bootstrapped with [monots].
-
-[monots]: https://github.com/monots/monots
