--- packages/@remirror/extension-base-keymap/src/index.ts
@@ -1,1 +1,0 @@
-export * from './base-keymap-extension';
