--- packages/remirror/extension/base-keymap/package.json
@@ -1,8 +1,0 @@
-{
-  "main": "dist/remirror.cjs.js",
-  "module": "dist/remirror.esm.js",
-  "types": "dist/remirror.cjs.d.ts",
-  "preconstruct": {
-    "source": "../../src/extension/base-keymap"
-  }
-}
