+++ packages/remirror/package.json
@@ -41,7 +41,6 @@
     "@remirror/dom": "^1.0.0-next.15",
     "@remirror/extension-annotation": "^0.0.0",
     "@remirror/extension-auto-link": "^1.0.0-next.15",
-    "@remirror/extension-base-keymap": "^1.0.0-next.15",
     "@remirror/extension-bidi": "^1.0.0-next.15",
     "@remirror/extension-blockquote": "^1.0.0-next.15",
     "@remirror/extension-bold": "^1.0.0-next.15",
@@ -129,7 +128,6 @@
       "core/utils",
       "dom",
       "extension/auto-link",
-      "extension/base-keymap",
       "extension/blockquote",
       "extension/bold",
       "extension/bidi",
