--- packages/remirror/src/extension/base-keymap.ts
@@ -1,1 +1,0 @@
-export * from '@remirror/extension-base-keymap';
