+++ support/deprecated/extension-base-keymap/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@remirror/extension-base-keymap",
+  "version": "1.0.0-next.13",
+  "description": "DEPRECATED: Add familiar keyboard shortcuts to your beautiful remirror editor.",
+  "keywords": [
+    "remirror",
+    "extension",
+    "keymaps",
+    "keyboard",
+    "shortcut",
+    "productivity"
+  ],
+  "homepage": "https://github.com/remirror/remirror/tree/HEAD/packages/@remirror/extension-base-keymap",
+  "repository": "https://github.com/remirror/remirror/tree/HEAD/support/deprecated/extension-base-keymap",
+  "license": "MIT",
+  "contributors": [
+    "Ifiok Jr. <ifiokotung@gmail.com>"
+  ],
+  "publishConfig": {
+    "access": "public"
+  }
+}
