+++ support/deprecated/extension-base-keymap/readme.md
@@ -1,0 +1,4 @@
+# @remirror/extension-base-keymap
+
+Deprecated - this functionality is now one of the `builtin` extensions. You can go ahead and remove
+this. Everything should work as before if you're on the latest version of `remirror`.
